+++ framework/sdk-parent/frontend/src/directive/index.js
@@ -1,4 +1,5 @@
 import ClickOutside from "element-ui/src/utils/clickoutside";
+import CustomClickOutside from "./click-outside/custom-click-outside";
 import Permission from "./permission";
 import Drag from "./drag";
 import directive from './loading/directive';
@@ -48,25 +49,7 @@ export default {
     Vue.directive('bottom-to-top-drag', Drag.bottom2TopDrag);
     Vue.directive('top-bottom-to-drag', Drag.top2BottomDrag);
     Vue.directive('vertical-drag', Drag.verticalDrag);
-    Vue.directive('my-click-outside', {
-      bind(el, binding, vnode) {
-        function clickHandler(e) {
-          if (el.contains(e.target)) {
-            return false;
-          }
-          if (binding.expression) {
-            binding.value(e);
-          }
-        }
-        el.__vueClickOutside__ = clickHandler;
-        document.addEventListener('click', clickHandler);
-      },
-      update() {},
-      unbind(el, binding) {
-        document.removeEventListener('click', el.__vueClickOutside__);
-        delete el.__vueClickOutside__;
-      },
-    });
+    Vue.directive('custom-click-outside', CustomClickOutside);
     Vue.use(directive);
     Vue.prototype.$loading = service;
   },
